requirements.txt
accelerate==1.10.1
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
diffusers==0.35.1
filelock==3.19.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.6.0
networkx==3.5
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.32.1
psutil==7.1.0
pyarrow==21.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rpds-py==0.27.1
safetensors==0.6.2
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0+cu128
torchvision==0.23.0+cu128
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
zipp==3.23.0
